requirements.txt
# Backend e APIs
fastapi==0.95.2
uvicorn==0.22.0
requests==2.31.0
python-multipart==0.0.6

# Redes Sociais
tweepy==4.14.0

# Processamento de Dados
pandas==1.5.3
numpy==1.24.3

# Dashboard e Visualização
streamlit==1.24.0
matplotlib==3.7.1
seaborn==0.12.2
wordcloud==1.9.2

# OCR e Validação de Documentos
pytesseract==0.3.10
Pillow==9.5.0

# Variáveis de Ambiente
python-dotenv==1.0.0